convertdbf2rdata.r has been modified to create one rds file from all the relevant SIP DBF files. The function is convert_1install_dbf_to_rds().
The raw data from SIP (stock investor pro) comes in a series of DBF files.  The Sample of DBF files.html lists all the DBFs for the first install.  The first part shows a few rows and fields for each DBF.  The second part lists all the fields for each dbf and the % of NA values.  There are R scripts within convertdbf2rdata.r to gather the info on the DBFs.